Moved statistics types and calcs to DistribCalcStuff.scala
