+++ grvx_mvn/grvx_zpol/src/main/scala/fun/gravax/zdynamo/GenTagStuff.scala
@@ -128,7 +128,7 @@ trait GenTagNumData {
 									val nextGenSt = GenSt(Some(nextPRec), nextAbsIdx, 1, nxtLevNum, nxtMaxKids)
 									(nextGenSt, mParQ)
 								} else {
-									// Parent says no kids so nothing to do here.
+									// Parent says no myKids so nothing to do here.
 									(noGenSt, mParQ)
 								}
 							}
